requirements.txt
# 1st activate the virtual env using venv\Scripts\activate

requests
pandas
python-dotenv
hopsworks==4.2.*
confluent-kafka
numpy
pyarrow
pymongo 
mlflow 
scikit-learn 
xgboost 
lightgbm
certifi   